second commit

+++ python/add.py
@@ -1,4 +1,4 @@
 num1 = input("Enter the first number: ")
 num2 = input("Enter the second number: ")
-sum_result = int(num1) + int(num2)
+sum_result = float(num1) + float(num2)
 print("Sum of first number and second number is:", sum_result)
